Created a star that rotates with mouse
However, transforming the projection and view matrices at runtime
is a lot harder that I thought it was. Going to continue with the
tutorials on open.gl and learnopengl.com on Sunday but will consult
the packt guide next week as I remember seeing a user controlled
camera in there.