Page 48 example code finished
Had trouble for about an hour or two tring to implement Angel's vector
library. However, I managed to fix this by referencing the code for
tessellation in Opengl: A Primer. I think using the primer for reference
for code but following the code from Interactive Computer Graphics will
be more beneficial for me since the latter explains the theory far more
thoroughly whilst the primer is lacking in this regard.