Completed Sierpinski.cpp ?
Am finding the interactive computer graphics guide very informative. I
think an understanding in the theory was what I needed when I started to
stray from the open.gl and learnopengl.com projects.
Going to consult the OpenGL programming guide to see if it has any
matierials on the process of compiling shaders from .glsl files: the one missing part of
my current Sierpinski project and a process that has caused me a lot of
greif across my entire process of learning opengl. The one cited at the end of Interactive
Computer Graphics uses the problematic libraries again which is a pain.
Opengl A primer is partially outliving its use at this point as it uses
deprecated functions and techniques. Looking forward to running through
it once I'm at the end of Interactive Computer Graphics.