Am successfully rendering triangle?
Scrapped the JoeyDeVries shader class in the end and ended up going
with the OGLProgGuide one as it was easier to fix the errors.
Ended up resorting to storing my vertices in a VBO as opposed to a shader,
but my fragment still isn't rendering at all: showing up as black for
nvidia card and white for AMD vega. I think I'll properly learn how
to do those later in the sb7.