Looked further into bresenham's lines
And created a project in visual studio so I could run through my
alg step by step. Did a lot of diagrams on paper for this (09/03/2020 - 10/03/2020)
and reverted the project to a console app for this commit. I don't know how my program
deals with 0 intersects but I'll cross that bridge when I come to it.
There shouldn't be a problem since C (change in x * zero intersect) is
cancelled out.